+++ crate/agent/src/main.rs
@@ -1,4 +1,5 @@
 use anyhow::Result;
+use cosmian_vm_agent::snapshot;
 use gethostname::gethostname;
 
 use actix_web::middleware::Logger;
@@ -41,12 +42,20 @@ async fn main() -> Result<()> {
     // First startup: initialize the agent
     initialize_agent(&conf)?;
 
+    // Background worker relating to the snapshot processing
+    tracing::info!("Starting the snapshot worker...");
+    let (snapshot_worker, snapshot_worker_handle, snapshot_worker_cancel) =
+        snapshot::init_snapshot_worker(conf.agent.tpm_device.clone());
+
     // Start REST server thread
-    tracing::info!("Starting server on {host}:{port}...");
+    tracing::info!("Starting Cosmian VM Agent on {host}:{port}...");
     HttpServer::new(move || {
         App::new()
             .wrap(Logger::default())
-            .configure(cosmian_vm_agent::config(conf.clone()))
+            .configure(cosmian_vm_agent::config(
+                conf.clone(),
+                snapshot_worker.clone(),
+            ))
     })
     .bind_rustls(
         format!("{host}:{port}"),
@@ -55,6 +64,14 @@ async fn main() -> Result<()> {
     .run()
     .await?;
 
+    tracing::info!("Stopping the snapshot worker...");
+    // signal snapshot worker to stop running
+    snapshot_worker_cancel.cancel();
+
+    // wait for the snapshot worker to exit its loop gracefully
+    snapshot_worker_handle.await.unwrap();
+
+    tracing::info!("Cosmian VM Agent successfully shutdown gracefully FFF");
     Ok(())
 }